Implemented Presset Periods for Datepicker
